4.06_test_weather_table

+++ Sunshine/app/src/main/java/com/freakybyte/sunshine/data/WeatherDbHelper.java
@@ -16,7 +16,7 @@ public class WeatherDbHelper extends SQLiteOpenHelper {
 
     // If you change the database schema, you must increment the database version.
     private static final int DATABASE_VERSION = 2;
-    public static final int FAIL_DB_MODIFY = -1;
+    public static final long FAIL_DB_MODIFY = -1L;
 
     static final String DATABASE_NAME = "weather.db";
